feat: viewing old note, disable update, delete btn

+++ views/noteView.view.php
@@ -19,17 +19,18 @@ $keywords = explode(", ", $keywords);
   <div class="ml-auto w-2/3 bg-white p-4">
     <div class="p-8 bg-white rounded-lg shadow-lg max-w-4xl mx-auto mt-8 space-y-12">
 
-      <!-- Top Right Buttons (Update & Delete) -->
-      <div class="flex justify-end space-x-4 mb-6">
-        <a href="/notes/update/<?= $note->id ?>" class="bg-green-500 text-white px-4 py-2 rounded-lg hover:bg-green-600 transition duration-300">
-          Update
-        </a>
-        <a href="/notes/delete/<?= $note->id ?>" class="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600 transition duration-300"
-          onclick="return confirm('Are you sure you want to delete this?');">
-          Delete
-        </a>
-      </div>
-
+      <?php if (!$viewingOldNote): ?>
+        <!-- Top Right Buttons (Update & Delete) -->
+        <div class="flex justify-end space-x-4 mb-6">
+          <a href="/notes/update/<?= $note->id ?>" class="bg-green-500 text-white px-4 py-2 rounded-lg hover:bg-green-600 transition duration-300">
+            Update
+          </a>
+          <a href="/notes/delete/<?= $note->id ?>" class="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600 transition duration-300"
+            onclick="return confirm('Are you sure you want to delete this?');">
+            Delete
+          </a>
+        </div>
+      <?php endif ?>
       <!-- Note Title and Date-Time -->
       <div class="space-y-2">
         <h2 class="text-4xl font-bold text-gray-800"><?= $note->title ?></h2>
